requirements.txt
# Core dependencies for stock data and analysis
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
python-dotenv
# HTTP requests and API integration
requests==2.31.0
urllib3==2.0.7

# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
apscheduler==3.10.4

# Email notifications
smtplib-ssl==1.3

# Scheduling and time handling
schedule==1.2.0
pytz==2023.3

# Environment variable management
python-dotenv==1.0.0

# Logging and monitoring
colorlog==6.7.0

# Optional: Telegram bot integration
python-telegram-bot==20.7

# Optional: Broker API integrations
smartapi-python==1.3.0  # Angel One API
upstox-client==2.0.0     # Upstox API

# Optional: Enhanced data analysis
ta==0.10.2              # Technical analysis library
matplotlib==3.7.2       # For local plotting/analysis
seaborn==0.12.2         # Enhanced visualization

# Development and testing (commented out for production)
# pytest==7.4.2
# pytest-cov==4.1.0
# black==23.7.0
# flake8==6.0.0
